api/news-handler/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
aiokafka==0.12.0
pydantic==2.7.1
kafka-python==2.1.2
aiocouch==1.2.0  
